Conflicts:
	src/main/java/br/com/ceos/controller/GrupoUsuarioCadastroController.java
	src/main/java/br/com/ceos/controller/GrupoUsuarioListaController.java
	src/main/java/br/com/ceos/controller/PrincipalController.java

+++ src/main/java/br/com/ceos/entity/GrupoUsuario.java
@@ -24,7 +24,7 @@ public class GrupoUsuario extends EntidadeBase implements Serializable {
   @Id
   @Column(name = "ID")
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Getter @Setter private int id;
+  @Getter @Setter private Long id;
 
   @Column(name = "DESCRICAO")
   @Getter @Setter private String descricao;
